added dependencies and requirements.txt
