+++ scripts/presigned_url_scripts/requirements.txt
@@ -1,0 +1,11 @@
+boto3==1.26.131
+botocore==1.29.131
+jmespath==1.0.1
+numpy==1.24.3
+pandas==2.0.1
+python-dateutil==2.8.2
+pytz==2023.3
+s3transfer==0.6.1
+six==1.16.0
+tzdata==2023.3
+urllib3==1.26.15
